refactor(invites-graph): decompose the 2.5k-loc InvitesGraph monolith in place (TASK-21854)
data, derived-memo, interaction, fetch, rendering, force and camera
concerns split into dedicated hooks; three jsx subviews extracted;
effect registration order kept byte-identical. index.tsx stays the
composition root with the same export and props. zero behavior change.
